dom/base/test/test_postMessages_workers.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for postMessages cloning/transferring objects</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script src="common_postMessages.js"></script>
</head>

<body>
<input id="fileList" type="file"></input>
<script type="application/javascript">

// PostMessage for Workers
tests.push(function test_workers() {
  info("Testing Workers");

  var resolve;

  var w = new Worker('worker_postMessages.js');
  w.postMessage('workers');
  w.onmessage = function(e) {
    is(e.data, 'ok', "Worker ready!");

    w.onmessage = function(e1) {
      if (!resolve) {
        ok(false, "Unexpected message!");
        return;
      }

      let tmp = resolve;
      resolve = null;
      tmp({ data: e1.data, ports: e1.ports, error: e1.data === "error" });
    }

    runTests({
      clonableObjectsEveryWhere: true,
      clonableObjectsSameProcess: true,
      transferableObjects: true,
      send(what, ports) {
        return new Promise(function(r, rr) {
          resolve = r;
          try {
            w.postMessage(what, ports);
          } catch(ex) {
            resolve = null;
            rr();
          }
        });
      },

      finished() {
        w.terminate();
        next();
      }
    });
  }
});

// PostMessage for SharedWorkers
tests.push(function test_sharedWorkers() {
  info("Testing SharedWorkers");

  var resolve;

  var w = new SharedWorker('worker_postMessages.js');
  w.port.postMessage('sharedworkers');
  w.port.onmessage = function(e) {
    is(e.data, 'ok', "Worker ready!");

    w.port.onmessage = function(e1) {
      if (!resolve) {
        ok(false, "Unexpected message!");
        return;
      }

      let tmp = resolve;
      resolve = null;
      tmp({ data: e1.data, ports: e1.ports, error: e1.data === "error" });
    }

    runTests({
      clonableObjectsEveryWhere: true,
      clonableObjectsSameProcess: false,
      transferableObjects: true,
      send(what, ports) {
        return new Promise(function(r, rr) {
          resolve = r;
          try {
            w.port.postMessage(what, ports);
          } catch(ex) {
            resolve = null;
            rr();
          }
        });
      },

      finished() {
        w.port.postMessage("terminate");
        next();
      }
    });
  }
});

SimpleTest.waitForExplicitFinish();
next();

</script>
</body>
</html>
